renamed route to quickseach and now use the apito get weather data on quicksearch instead of the scraper
