+++ lib/scheduler/wbgtAlert.js
@@ -1,6 +1,6 @@
 const cron = require('node-cron');
-const { Alert, Wbgt, User } = require('../models')
-const getWbgtInfo = require('../scraper/scraper');
+const { Alert, Wbgt } = require('../models')
+const getWbgtInfo = require('../wgbt/getWbgtInfo');
 
 module.exports = () => {  
   cron.schedule('0 */10 * * * *', async () => {
@@ -33,7 +33,7 @@ module.exports = () => {
         time: alertInfo.dateTimeInfo.time,
         solarRadiance: alertInfo.wgbtInfo.solarRadiance,
         directWBGT: alertInfo.wgbtInfo.directWBGT,
-        shadedWBGT: alertInfo.wgbtInfo.shadeWBGT,
+        shadedWBGT: alertInfo.wgbtInfo.shadedWBGT,
         heatIndex: alertInfo.wgbtInfo.heatIndex
       }
