+++ lib/server.js
@@ -4,14 +4,14 @@ const app = express();
 const cors = require('cors');
 app.use(cors());
 app.use(express.json());
-const scraperRoute = require('./routes/scraper');
+const quickSearchRoute = require('./routes/quickSearch');
 const alertRoute = require('./routes/alert');
 const userRoute = require('./routes/user');
 const wgbtRoute = require('./routes/wgbt');
 const weatherApi = require('./routes/weatherApi');
 require('./scheduler/wbgtAlert')();
 
-app.use(scraperRoute);
+app.use(quickSearchRoute);
 app.use(alertRoute);
 app.use(userRoute);
 app.use(wgbtRoute);
