+++ lib/wgbt/getWbgtInfo.js
@@ -1,42 +1,31 @@
 'use strict';
-const express = require('express');
-const router = express.Router();
 const axios = require('axios');
-const cheerio = require('cheerio');
 const puppeteer = require('puppeteer');
 
-router.get('/scraper/', async (req, res) => {
-  var currentdate = new Date(); 
-  var time = `${currentdate.getHours()}:${currentdate.getMinutes()}`
-  var date = `${currentdate.getMonth()+1}/${currentdate.getDate()}/${currentdate.getFullYear()}`
-
-  const {lat, lon} = req.query
-  const url = `https://forecast.weather.gov/MapClick.php?lat=${lat}&lon=${lon}`;
-  try {
-    const response = await axios.get(url);
-    const $ = cheerio.load(response.data);
-    const temperature = $('p.myforecast-current-lrg').text();
-
-    // console.log(weatherRows.children)
-    const weatherRows = []
-    $('tr > td').each((_idx, el) => {
-      const info = $(el).text()
-      weatherRows.push(info)
-    })
-    console.log(weatherRows)
-    let information = {
-      weatherInfo:{
-        'temperature': temperature.slice(0,2),
-        'humidity': weatherRows[1].slice(0,2).replace('%',''),
-        'windspeed': weatherRows[3].replace(/\D/g, ""),
-        'barometer': weatherRows[5].slice(0,5)
-      },
-      dateTimeInfo: {
-        date: date,
-        time: time
-      }
+const getWbgtInfo = async (lat, lon) => {
+  const getWeather = await axios.get(`https://api.openweathermap.org/data/2.5/weather?lat=${lat}&lon=${lon}&appid=698c67f9881bc777269a22e04b573368`)
+  console.log(getWeather.data, 'this is the weather data')
+  let unix_timestamp = getWeather.data.dt
+  var newDate = new Date(unix_timestamp * 1000);
+  console.log(newDate, 'this is the newDate')
+  var date = `${newDate.getMonth()+1}/${newDate.getDate()}/${newDate.getFullYear()}`
+  var time = newDate.toLocaleTimeString("it-IT").slice(0,5);
+  const windspeed = Math.round(((getWeather.data.wind.speed * 2.237) + Number.EPSILON) * 100) / 100
+  const temperature = Math.round((((getWeather.data.main.temp - 273.15)*1.8)+32 + Number.EPSILON) * 100) / 100
+  const barometer = Math.round(((getWeather.data.main.pressure /  33.86389) + Number.EPSILON) * 100) / 100
+  let information = {
+    weatherInfo:{
+      'temperature': temperature,
+      'humidity': getWeather.data.main.humidity,
+      'windspeed': windspeed,
+      'barometer': barometer
+    },
+    dateTimeInfo: {
+      date: date,
+      time: time
     }
-
+  }
+  console.log(information,'this is the information')
     const browser = await puppeteer.launch();
     const page = await browser.newPage();
     await page.goto('https://www.osha.gov/heat-exposure/wbgt-calculator');
@@ -65,12 +54,8 @@ router.get('/scraper/', async (req, res) => {
         heatIndex: heatIndex.split('/')[0].replace(/\D/g, "")
       }
     }
-    console.log(information)
-    res.send(information);
-    
-  } catch (e) {
-    console.error(e);
-  }
-});
+    console.log(information);
+    return information;
+}
 
-module.exports = router;
+module.exports = getWbgtInfo;
